+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 ----------
 Added
 ^^^^^
+- Automatic creation of PREMIS events when adding imported metadata to structural map
 - Ability to create structural maps from directory structures
 
 0.1.0 - 2024-03-27